Refactor components out to separate files
